refactor(channel): split cli_message.go (4038 lines) into 11 focused files
Original cli_message.go was a monolithic 4038-line file with 71 functions
spanning 12 functional domains. Split into logically cohesive modules:

  cli_tab.go         (126 lines) — Tab completion (command + @file path)
  cli_inbound.go     (207 lines) — Message sending / input processing
  cli_cache.go       (334 lines) — Render cache management (invalidate, rebuild, merge)
  cli_slash.go       (317 lines) — Slash command handler (/help, /clear, etc.)
  cli_agent_msg.go   (432 lines) — Agent message handler (streaming, turn state)
  cli_progress.go    (820 lines) — Progress block rendering (iteration, tools, reasoning)
  cli_msg_render.go  (567 lines) — Single message rendering (by role: user/assistant/system)
  cli_viewport.go    (346 lines) — Viewport content management + streaming fast path
  cli_search.go      (163 lines) — Message search, fold, typewriter tick
  cli_tool_render.go (407 lines) — Tool body rendering (Read/Shell/Grep/Glob)
  cli_diff.go        (396 lines) — Diff rendering with syntax highlighting

cli_message.go retained as a 15-line index comment pointing to the new files.

No behavioral changes — pure structural refactoring within the same package.
